Move power bloq to basic_gates
