+++ qualtran/bloqs/bookkeeping/__init__.py
@@ -19,5 +19,4 @@ from qualtran.bloqs.bookkeeping.cast import Cast
 from qualtran.bloqs.bookkeeping.free import Free
 from qualtran.bloqs.bookkeeping.join import Join
 from qualtran.bloqs.bookkeeping.partition import Partition
-from qualtran.bloqs.bookkeeping.power import Power
 from qualtran.bloqs.bookkeeping.split import Split
